type improvements
