+++ src/index.ts
@@ -1,16 +1,10 @@
-export {
-  LayerComposerStyles,
-  LayerComposerOptions,
-  GeneratorStyles,
-  Generator,
-  GeneratorConfig,
-} from './layer-composer/types'
+export { LayerComposerStyles, LayerComposerOptions, GeneratorStyles } from './layer-composer/types'
+export { Type, Generator, GeneratorConfig } from './layer-composer/generators/types'
 
 export { default, DEFAULT_CONFIG } from './layer-composer'
 
 export {
   default as defaultGenerators,
-  TYPES,
   HEATMAP_GEOM_TYPES,
   HEATMAP_COLOR_RAMPS,
 } from './layer-composer/generators'
