+++ src/layer-composer/generators/track/track.ts
@@ -1,14 +1,11 @@
 import { scaleLinear, scalePow } from 'd3-scale'
 import { FeatureCollection, LineString } from 'geojson'
 import memoizeOne from 'memoize-one'
-import { GeneratorConfig } from 'layer-composer/types'
-// TODO custom "augmented" GeoJSON type?
-// see https://github.com/yagajs/generic-geojson/blob/master/index.d.ts
+import { Group } from '../../types'
+import { Type, TrackGeneratorConfig } from '../types'
 import filterGeoJSONByTimerange from './filterGeoJSONByTimerange'
 import { simplifyTrack } from './simplify-track'
 import { memoizeByLayerId, memoizeCache } from '../../utils'
-
-export const TRACK_TYPE = 'TRACK'
 
 const mapZoomToMinPosΔ = (zoomLoadLevel: number) => {
   // first normalize and invert z level
@@ -30,12 +27,6 @@ const mapZoomToMinPosΔ = (zoomLoadLevel: number) => {
   return minPosΔ
 }
 
-export interface TrackGeneratorConfig extends GeneratorConfig {
-  data: FeatureCollection
-  simplify?: boolean
-  color?: string
-}
-
 const simplifyTrackWithZoomLevel = (
   data: FeatureCollection,
   zoomLoadLevel: number
@@ -54,7 +45,7 @@ const filterByTimerange = (data: FeatureCollection, start: string, end: string) 
 }
 
 class TrackGenerator {
-  type = TRACK_TYPE
+  type = Type.Track
 
   _getStyleSources = (config: TrackGeneratorConfig) => {
     const defaultGeoJSON: FeatureCollection = {
@@ -88,6 +79,9 @@ class TrackGenerator {
       source: config.id,
       layout: {},
       paint: { 'line-color': config.color || 'hsl(100, 100%, 55%)' },
+      metadata: {
+        group: Group.Track,
+      },
     }
     return [layer]
   }
